Moved exceptions into sub-namespace.
