+++ BNA/Values/Value.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using BNA.Exceptions;
 
 namespace BNA.Values
 {
